+++ examples/hello_world.rs
@@ -4,7 +4,7 @@ fn main() {
     let mut ctx = ClipboardContext::new().unwrap();
 
     let msg = "Hello, world!";
-    ctx.set_contents(msg).unwrap();
+    ctx.set_contents(msg.to_owned()).unwrap();
 
     let content = ctx.get_contents().unwrap();