+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Use `String` in `ClipboardProvider::set_contents` for trait to be *object-safe*
+
 ## 0.9.0
 
 - Bump minimum supported Rust version to `1.65.0`